Fix Garbage Disposal
Garbage disposal deletes all items in the trash can once enough items have piled up.

+++ Cooking VR/Assets/Garbage_Disposal.cs
@@ -9,12 +9,18 @@ public class Garbage_Disposal : MonoBehaviour
     [SerializeField]
     BoxCollider entryCollider;
 
+    [SerializeField]
+    float fullCheckTime;
+    [SerializeField]
+    float fullCheckTimeInterval;
+
+
     // Start is called before the first frame update
     void Start()
     {
         garbage = new List<GameObject>();
-       
-     
+
+        fullCheckTimeInterval = 3;
     }
 
     // Update is called once per frame
@@ -25,24 +31,43 @@ public class Garbage_Disposal : MonoBehaviour
 
     private void OnTriggerEnter(Collider other)
     {
-        garbage.Add(other.gameObject);
-        Debug.Log("Trash has entered the field");
+        if(!garbage.Contains(other.gameObject))
+        {
+            garbage.Add(other.gameObject);
+            //Debug.Log("Trash has entered the field");
+            Debug.Log($"{other.gameObject.name} has entered the trash can");
+            fullCheckTime = Time.time + fullCheckTimeInterval;
+        }
+            
     }
 
     private void OnTriggerStay(Collider other)
     {
-        Debug.Log("KNOW YOUR PLACE TRASH!");
-        DeleteItem(true);
+        foreach(GameObject trash in garbage)
+        {
+            if(other.gameObject.Equals(trash))
+            {
+                if (Time.time >= fullCheckTime)
+                {
+                    Debug.Log("KNOW YOUR PLACE TRASH!");
+                    DeleteItem(true);
+                }
+                    
+            }
+        }
+        
     }
 
     void DeleteItem(bool isFull = false)
     {
        if(isFull == true)
         {
-            foreach (GameObject trash in garbage)
+            //foreach (GameObject trash in garbage)
+            for(int i = garbage.Count -1; i >= 0; i--)
             {
-                garbage.Remove(trash);
-                Destroy(trash);
+                Debug.Log($"{garbage[i].name} has been deleted!");
+                Destroy(garbage[i]);
+                garbage.Remove(garbage[i]);
             }
         }
     }
